resolve PR comment: separate StartInvestigationButton logic into separate hook
