+++ public/app/features/alerting/unified/triage/instance-details/StartInvestigationButton.tsx
@@ -1,24 +1,9 @@
-import { skipToken } from '@reduxjs/toolkit/query';
-import { useEffect, useMemo, useState } from 'react';
-
 import { type Labels } from '@grafana/data';
 import { Trans, t } from '@grafana/i18n';
-import { config } from '@grafana/runtime';
 import { Button, LinkButton, Spinner, Stack, Text, Tooltip } from '@grafana/ui';
-import { GrafanaAlertState, type GrafanaRuleDefinition } from 'app/types/unified-alerting-dto';
+import { type GrafanaAlertState, type GrafanaRuleDefinition } from 'app/types/unified-alerting-dto';
 
-import {
-  ASSISTANT_INVESTIGATION_POLL_INTERVAL_MS,
-  type StartInvestigationFromAlertRequest,
-  assistantApi,
-  getAssistantInvestigationUrl,
-  isAssistantInvestigationCompleted,
-  isAssistantInvestigationFailed,
-  isAssistantInvestigationTerminal,
-} from '../../api/assistantApi';
-import { usePluginBridge } from '../../hooks/usePluginBridge';
-import { SupportedPlugin } from '../../types/pluginBridges';
-import { createAbsoluteUrl } from '../../utils/url';
+import { useStartInvestigation } from './useStartInvestigation';
 
 interface StartInvestigationButtonProps {
   instanceLabels: Labels;
@@ -27,22 +12,12 @@ interface StartInvestigationButtonProps {
   alertState?: GrafanaAlertState | null;
 }
 
-/** True when both product toggles for the manual investigation entry point are on. */
-export function isManualAssistantInvestigationEnabled(): boolean {
-  return Boolean(
-    config.featureToggles.alertingEnrichmentAssistantInvestigations &&
-      config.featureToggles.alertingManualAssistantInvestigation
-  );
-}
-
 /**
  * POC: lets a responder manually start (or open) a Grafana Assistant investigation
  * for the firing alert instance they are looking at. Mirrors the three states of the
  * incident-sidebar treatment (not started / running / completed).
  *
- * The Assistant dedups by alert group (`groupLabels`), so pressing this repeatedly
- * opens the existing investigation rather than creating duplicates. Reopening the
- * drawer looks up that same link and polls until the report completes.
+ * State, API calls, and URL generation live in {@link useStartInvestigation}.
  */
 export function StartInvestigationButton({
   instanceLabels,
@@ -50,329 +25,156 @@ export function StartInvestigationButton({
   rule,
   alertState,
 }: StartInvestigationButtonProps) {
-  const featureEnabled = isManualAssistantInvestigationEnabled();
-  const { installed } = usePluginBridge(SupportedPlugin.Assistant);
+  const view = useStartInvestigation({ instanceLabels, commonLabels, rule, alertState });
 
-  // Stable identity for RTK Query cache keys — omit startsAt/status/name/generatorURL
-  // (those are attached only on create). Wait for rule identity when the instance has
-  // no labels, otherwise early Start/lookup can hash a different group key once the
-  // rule arrives.
-  const hasStableIdentity = Object.keys(instanceLabels).length > 0 || Boolean(rule?.uid) || Boolean(rule?.title);
+  switch (view.status) {
+    case 'hidden':
+      return null;
 
-  const requestBody = useMemo(
-    () => (hasStableIdentity ? buildFromAlertRequest({ instanceLabels, commonLabels, rule }) : null),
-    // Labels are plain objects from parents; stringify keeps the body stable across
-    // equivalent re-renders so lookup does not thrash. alertState is intentionally
-    // excluded — firing↔resolved must not change the cache key mid-drawer.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [JSON.stringify(instanceLabels), JSON.stringify(commonLabels), rule?.uid, rule?.title, hasStableIdentity]
-  );
+    case 'waitingIdentity':
+      return (
+        <Tooltip
+          content={t(
+            'alerting.triage.instance-details-drawer.investigation-waiting-identity',
+            'Waiting for alert details before starting an investigation.'
+          )}
+        >
+          <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" disabled>
+            <Trans i18nKey="alerting.triage.instance-details-drawer.start-investigation">Start investigation</Trans>
+          </Button>
+        </Tooltip>
+      );
 
-  const requestBodyKey = useMemo(() => (requestBody ? JSON.stringify(requestBody) : ''), [requestBody]);
+    case 'lookingUp':
+      return (
+        <Stack direction="row" alignItems="center" gap={1}>
+          <Spinner size="sm" inline />
+          <Text variant="bodySmall" color="secondary">
+            <Trans i18nKey="alerting.triage.instance-details-drawer.investigation-checking">
+              Checking for investigation…
+            </Trans>
+          </Text>
+        </Stack>
+      );
 
-  const [startInvestigation, { isLoading, data, isError, reset, originalArgs }] =
-    assistantApi.useStartInvestigationFromAlertMutation();
+    case 'lookupError':
+      return (
+        <Tooltip
+          content={t(
+            'alerting.triage.instance-details-drawer.investigation-lookup-failed',
+            'Could not check for an existing investigation. Try again.'
+          )}
+        >
+          <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" onClick={view.onRetry}>
+            <Trans i18nKey="alerting.triage.instance-details-drawer.investigation-lookup-retry">Retry check</Trans>
+          </Button>
+        </Tooltip>
+      );
 
-  // Mutation result is shared RTK state. Only trust it for the current alert identity
-  // (create-only fields stripped via stableFromAlertRequest).
-  const mutationMatchesCurrent = useMemo(() => {
-    if (!originalArgs || !requestBodyKey) {
-      return false;
-    }
-    return JSON.stringify(stableFromAlertRequest(originalArgs)) === requestBodyKey;
-  }, [originalArgs, requestBodyKey]);
-
-  const startedInvestigation = mutationMatchesCurrent ? data : undefined;
-  const isStarting = mutationMatchesCurrent && isLoading;
-  const isStartError = mutationMatchesCurrent && isError;
-
-  // Drop stale mutation state when the drawer switches instances.
-  useEffect(() => {
-    if (!mutationMatchesCurrent) {
-      reset();
-    }
-  }, [mutationMatchesCurrent, reset]);
-
-  const {
-    data: lookedUpInvestigation,
-    isLoading: isLookingUp,
-    isError: isLookupError,
-    refetch: refetchLookup,
-  } = assistantApi.useLookupInvestigationFromAlertQuery(requestBody ?? skipToken, {
-    skip: !featureEnabled || !installed,
-  });
-
-  const knownId = startedInvestigation?.id ?? lookedUpInvestigation?.id;
-  const [shouldPoll, setShouldPoll] = useState(false);
-
-  const {
-    data: polledInvestigation,
-    isError: isPollError,
-    refetch: refetchPoll,
-  } = assistantApi.useGetAssistantInvestigationQuery(knownId ?? '', {
-    skip: !featureEnabled || !installed || !knownId,
-    pollingInterval: shouldPoll ? ASSISTANT_INVESTIGATION_POLL_INTERVAL_MS : 0,
-  });
-
-  // Prefer the create/retry snapshot until poll has data for that same investigation
-  // id. Do not prefer mutation over poll for the same id — that would freeze the UI
-  // on the initial pending/in_progress snapshot after the report completes. Retry
-  // after failure creates a new id, so the id mismatch branch still covers that case.
-  const investigation = useMemo(() => {
-    if (startedInvestigation && (!polledInvestigation || polledInvestigation.id !== startedInvestigation.id)) {
-      return startedInvestigation;
-    }
-    return polledInvestigation ?? startedInvestigation ?? lookedUpInvestigation ?? undefined;
-  }, [polledInvestigation, startedInvestigation, lookedUpInvestigation]);
-
-  const investigationFailed = isAssistantInvestigationFailed(investigation?.state);
-
-  useEffect(() => {
-    if (!knownId || isPollError) {
-      // Stop on poll failure so we don't keep spinning "Generating…" with no recovery.
-      setShouldPoll(false);
-      return;
-    }
-    // Keep polling until a terminal state — unknown/paused must not freeze the UI.
-    setShouldPoll(!isAssistantInvestigationTerminal(investigation?.state ?? 'pending'));
-  }, [knownId, investigation?.state, isPollError]);
-
-  if (!featureEnabled || !installed) {
-    return null;
-  }
-
-  const handleStart = () => {
-    if (!requestBody) {
-      return;
-    }
-    const startsAt = new Date().toISOString();
-    const status = alertState === GrafanaAlertState.Normal ? 'resolved' : 'firing';
-    const generatorURL = rule?.uid ? createAbsoluteUrl(`/alerting/grafana/${rule.uid}/view`) : undefined;
-    startInvestigation({
-      ...requestBody,
-      name: rule?.title,
-      alerts: requestBody.alerts.map((alert) => ({ ...alert, startsAt, status, generatorURL })),
-    });
-  };
-
-  // No stable group key yet (empty labels and rule still loading) — don't pretend
-  // we're looking up; Start stays disabled until identity is ready.
-  if (!hasStableIdentity) {
-    return (
-      <Tooltip
-        content={t(
-          'alerting.triage.instance-details-drawer.investigation-waiting-identity',
-          'Waiting for alert details before starting an investigation.'
-        )}
-      >
-        <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" disabled>
-          <Trans i18nKey="alerting.triage.instance-details-drawer.start-investigation">Start investigation</Trans>
-        </Button>
-      </Tooltip>
-    );
-  }
-
-  if (isLookingUp && !investigation) {
-    return (
-      <Stack direction="row" alignItems="center" gap={1}>
-        <Spinner size="sm" inline />
-        <Text variant="bodySmall" color="secondary">
-          <Trans i18nKey="alerting.triage.instance-details-drawer.investigation-checking">
-            Checking for investigation…
-          </Trans>
-        </Text>
-      </Stack>
-    );
-  }
-
-  // Non-404 lookup failures must not look like "no investigation" (would offer Start incorrectly).
-  if (isLookupError && !investigation) {
-    return (
-      <Tooltip
-        content={t(
-          'alerting.triage.instance-details-drawer.investigation-lookup-failed',
-          'Could not check for an existing investigation. Try again.'
-        )}
-      >
-        <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" onClick={() => refetchLookup()}>
-          <Trans i18nKey="alerting.triage.instance-details-drawer.investigation-lookup-retry">Retry check</Trans>
-        </Button>
-      </Tooltip>
-    );
-  }
-
-  if (investigation && isAssistantInvestigationCompleted(investigation.state)) {
-    return (
-      <LinkButton
-        icon="file-alt"
-        variant="primary"
-        fill="text"
-        size="sm"
-        href={getAssistantInvestigationUrl(investigation.id)}
-        target="_blank"
-        rel="noopener noreferrer"
-      >
-        <Trans i18nKey="alerting.triage.instance-details-drawer.open-investigation">
-          Open full investigation report
-        </Trans>
-      </LinkButton>
-    );
-  }
-
-  if (isStarting) {
-    return (
-      <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" disabled>
-        <Trans i18nKey="alerting.triage.instance-details-drawer.investigation-starting">Starting…</Trans>
-      </Button>
-    );
-  }
-
-  if (isStartError) {
-    return (
-      <Tooltip
-        content={t(
-          'alerting.triage.instance-details-drawer.investigation-start-failed',
-          'Could not start the investigation. Try again.'
-        )}
-      >
-        <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" onClick={handleStart}>
-          <Trans i18nKey="alerting.triage.instance-details-drawer.start-investigation">Start investigation</Trans>
-        </Button>
-      </Tooltip>
-    );
-  }
-
-  if (investigationFailed && investigation) {
-    return (
-      <Tooltip
-        content={t(
-          'alerting.triage.instance-details-drawer.investigation-report-failed',
-          'The investigation report failed or was cancelled. You can try again.'
-        )}
-      >
-        <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" onClick={handleStart}>
-          <Trans i18nKey="alerting.triage.instance-details-drawer.start-investigation">Start investigation</Trans>
-        </Button>
-      </Tooltip>
-    );
-  }
-
-  // Poll failures must not leave the generating spinner stuck on a stale non-terminal
-  // snapshot from lookup/start. Offer an explicit retry that resumes polling on success.
-  if (isPollError && knownId && !isAssistantInvestigationTerminal(investigation?.state)) {
-    return (
-      <Tooltip
-        content={t(
-          'alerting.triage.instance-details-drawer.investigation-poll-failed',
-          'Could not refresh investigation status. Try again.'
-        )}
-      >
-        <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" onClick={() => refetchPoll()}>
-          <Trans i18nKey="alerting.triage.instance-details-drawer.investigation-poll-retry">Retry status</Trans>
-        </Button>
-      </Tooltip>
-    );
-  }
-
-  // Active, paused, or any non-terminal/unknown state — keep generating UI while polling.
-  if (investigation && !isAssistantInvestigationTerminal(investigation.state)) {
-    return (
-      <Stack direction="row" alignItems="center" gap={1}>
-        <Spinner size="sm" inline />
-        <Text variant="bodySmall" color="secondary">
-          <Trans i18nKey="alerting.triage.instance-details-drawer.investigation-running">
-            Generating investigation report…
-          </Trans>
-        </Text>
+    case 'completed':
+    case 'open':
+      return (
         <LinkButton
-          icon="comment-alt"
+          icon="file-alt"
           variant="primary"
           fill="text"
           size="sm"
-          href={getAssistantInvestigationUrl(investigation.id)}
+          href={view.href}
           target="_blank"
           rel="noopener noreferrer"
         >
-          <Trans i18nKey="alerting.triage.instance-details-drawer.investigation-watch-live">
-            Watch live in assistant workspace
+          <Trans i18nKey="alerting.triage.instance-details-drawer.open-investigation">
+            Open full investigation report
           </Trans>
         </LinkButton>
-      </Stack>
-    );
+      );
+
+    case 'starting':
+      return (
+        <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" disabled>
+          <Trans i18nKey="alerting.triage.instance-details-drawer.investigation-starting">Starting…</Trans>
+        </Button>
+      );
+
+    case 'startError':
+      return (
+        <Tooltip
+          content={t(
+            'alerting.triage.instance-details-drawer.investigation-start-failed',
+            'Could not start the investigation. Try again.'
+          )}
+        >
+          <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" onClick={view.onStart}>
+            <Trans i18nKey="alerting.triage.instance-details-drawer.start-investigation">Start investigation</Trans>
+          </Button>
+        </Tooltip>
+      );
+
+    case 'reportFailed':
+      return (
+        <Tooltip
+          content={t(
+            'alerting.triage.instance-details-drawer.investigation-report-failed',
+            'The investigation report failed or was cancelled. You can try again.'
+          )}
+        >
+          <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" onClick={view.onStart}>
+            <Trans i18nKey="alerting.triage.instance-details-drawer.start-investigation">Start investigation</Trans>
+          </Button>
+        </Tooltip>
+      );
+
+    case 'pollError':
+      return (
+        <Tooltip
+          content={t(
+            'alerting.triage.instance-details-drawer.investigation-poll-failed',
+            'Could not refresh investigation status. Try again.'
+          )}
+        >
+          <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" onClick={view.onRetry}>
+            <Trans i18nKey="alerting.triage.instance-details-drawer.investigation-poll-retry">Retry status</Trans>
+          </Button>
+        </Tooltip>
+      );
+
+    case 'running':
+      return (
+        <Stack direction="row" alignItems="center" gap={1}>
+          <Spinner size="sm" inline />
+          <Text variant="bodySmall" color="secondary">
+            <Trans i18nKey="alerting.triage.instance-details-drawer.investigation-running">
+              Generating investigation report…
+            </Trans>
+          </Text>
+          <LinkButton
+            icon="comment-alt"
+            variant="primary"
+            fill="text"
+            size="sm"
+            href={view.href}
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            <Trans i18nKey="alerting.triage.instance-details-drawer.investigation-watch-live">
+              Watch live in assistant workspace
+            </Trans>
+          </LinkButton>
+        </Stack>
+      );
+
+    case 'idle':
+      return (
+        <Tooltip
+          content={t(
+            'alerting.triage.instance-details-drawer.start-investigation-tooltip',
+            'Assistant will draft it in your workspace'
+          )}
+        >
+          <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" onClick={view.onStart}>
+            <Trans i18nKey="alerting.triage.instance-details-drawer.start-investigation">Start investigation</Trans>
+          </Button>
+        </Tooltip>
+      );
   }
-
-  // Known investigation in an unexpected terminal state — open it; never offer Start.
-  if (investigation) {
-    return (
-      <LinkButton
-        icon="file-alt"
-        variant="primary"
-        fill="text"
-        size="sm"
-        href={getAssistantInvestigationUrl(investigation.id)}
-        target="_blank"
-        rel="noopener noreferrer"
-      >
-        <Trans i18nKey="alerting.triage.instance-details-drawer.open-investigation">
-          Open full investigation report
-        </Trans>
-      </LinkButton>
-    );
-  }
-
-  return (
-    <Tooltip
-      content={t(
-        'alerting.triage.instance-details-drawer.start-investigation-tooltip',
-        'Assistant will draft it in your workspace'
-      )}
-    >
-      <Button icon="ai-sparkle" variant="primary" fill="text" size="sm" onClick={handleStart}>
-        <Trans i18nKey="alerting.triage.instance-details-drawer.start-investigation">Start investigation</Trans>
-      </Button>
-    </Tooltip>
-  );
-}
-
-/**
- * Strips per-delivery / rule-metadata fields so create/lookup share one RTK cache
- * identity. startsAt, status, name, and generatorURL can appear after Start or when
- * the rule query resolves; group identity (labels) must not change mid-drawer.
- */
-export function stableFromAlertRequest(body: StartInvestigationFromAlertRequest): StartInvestigationFromAlertRequest {
-  const { name: _name, ...rest } = body;
-  return {
-    ...rest,
-    alerts: body.alerts.map(({ startsAt: _startsAt, status: _status, generatorURL: _generatorURL, ...alert }) => alert),
-  };
-}
-
-/** Exported for unit tests — builds the stable from-alert payload (no startsAt/status/name/generatorURL). */
-export function buildFromAlertRequest({
-  instanceLabels,
-  commonLabels,
-  rule,
-}: Omit<StartInvestigationButtonProps, 'alertState'>): StartInvestigationFromAlertRequest {
-  const externalURL = config.appUrl.replace(/\/$/, '');
-
-  // Stable alert-group identity for Assistant dedup/lookup. Prefer instance labels;
-  // when an instance has none, fall back to rule identity so reopen still finds the link.
-  const groupLabels: Record<string, string> =
-    Object.keys(instanceLabels).length > 0
-      ? { ...instanceLabels }
-      : {
-          ...(rule?.title ? { alertname: rule.title } : {}),
-          ...(rule?.uid ? { rule_uid: rule.uid } : {}),
-        };
-
-  return {
-    alerts: [
-      {
-        labels: instanceLabels,
-      },
-    ],
-    commonLabels,
-    groupLabels,
-    externalURL,
-  };
 }
