+++ public/app/features/alerting/unified/triage/instance-details/startInvestigationFromAlert.test.ts
@@ -1,6 +1,15 @@
 import { config } from '@grafana/runtime';
 
-import { buildFromAlertRequest, stableFromAlertRequest } from './StartInvestigationButton';
+import { stableFromAlertRequest } from '../../api/assistantApi';
+
+import {
+  buildFromAlertRequest,
+  getAssistantInvestigationUrl,
+  isAssistantInvestigationActive,
+  isAssistantInvestigationCompleted,
+  isAssistantInvestigationFailed,
+  isAssistantInvestigationTerminal,
+} from './startInvestigationFromAlert';
 
 describe('buildFromAlertRequest', () => {
   const originalAppUrl = config.appUrl;
@@ -88,5 +97,54 @@ describe('stableFromAlertRequest', () => {
     expect(afterStart.alerts[0].status).toBeUndefined();
     expect(afterStart.alerts[0].startsAt).toBeUndefined();
     expect(afterStart.alerts[0].generatorURL).toBeUndefined();
+  });
+});
+
+describe('investigation state helpers', () => {
+  describe('isAssistantInvestigationActive', () => {
+    it.each(['pending', 'running', 'in_progress', 'in-progress', 'paused'])('treats %s as active', (state) => {
+      expect(isAssistantInvestigationActive(state)).toBe(true);
+    });
+
+    it.each(['completed', 'failed', 'cancelled', 'canceled', undefined, ''])('treats %s as inactive', (state) => {
+      expect(isAssistantInvestigationActive(state)).toBe(false);
+    });
+  });
+
+  describe('isAssistantInvestigationCompleted', () => {
+    it('is true only for completed', () => {
+      expect(isAssistantInvestigationCompleted('completed')).toBe(true);
+      expect(isAssistantInvestigationCompleted('in_progress')).toBe(false);
+    });
+  });
+
+  describe('isAssistantInvestigationFailed', () => {
+    it.each(['failed', 'cancelled', 'canceled'])('treats %s as failed', (state) => {
+      expect(isAssistantInvestigationFailed(state)).toBe(true);
+    });
+
+    it('is false for other states', () => {
+      expect(isAssistantInvestigationFailed('completed')).toBe(false);
+      expect(isAssistantInvestigationFailed(undefined)).toBe(false);
+    });
+  });
+
+  describe('isAssistantInvestigationTerminal', () => {
+    it.each(['completed', 'failed', 'cancelled', 'canceled'])('treats %s as terminal', (state) => {
+      expect(isAssistantInvestigationTerminal(state)).toBe(true);
+    });
+
+    it.each(['pending', 'in_progress', 'paused', 'weird-future-state', undefined, ''])(
+      'treats %s as non-terminal',
+      (state) => {
+        expect(isAssistantInvestigationTerminal(state)).toBe(false);
+      }
+    );
+  });
+
+  describe('getAssistantInvestigationUrl', () => {
+    it('builds a plugin bridge URL with an encoded investigation id', () => {
+      expect(getAssistantInvestigationUrl('inv/123')).toBe('/a/grafana-assistant-app/investigations/inv%2F123');
+    });
   });
 });
